+++ bin/varnish-plus-cli.php
@@ -5,13 +5,15 @@ declare(strict_types=1);
 
 require __DIR__.'/../vendor/autoload.php';
 
-use App\Command\VclDeployCommand;
+use App\Command\VacDeployCommand;
+use App\Command\VarnishControllerDeployCommand;
 use App\Command\VclTwigCompileCommand;
 use Symfony\Component\Console\Application;
 
 $application = new Application('Varnish Plus CLI');
 
 $application->add(new VclTwigCompileCommand());
-$application->add(new VclDeployCommand());
+$application->add(new VarnishControllerDeployCommand());
+$application->add(new VacDeployCommand());
 
 $application->run();